FULL REWRITE: Build from source
Instead of just doing a simple apt install we now download the latest
source code and build Bind. This means we can upload the latest
version and not just what is available in the apt repo.

We also include all the extra configuration files available here to
highlight what they contain.