Build both "dev" and "stable" releases
ISC use the minor version number to define if this is a stable or a
development release, which means we should not treat a larger number
blindly as more recent.

Here we set up so that we build both the latest stable and
development versions, with the "latest" tag following the dev branch
and the sole major version tag following the stable branch.